+++ R/file_engine.R
@@ -8,8 +8,6 @@
 #' stored in the engine's preferences. A default set of preferences is stored in the package for use with KGX (BioLink-compatible) graphs (see https://github.com/biolink/kgx/blob/master/specification/kgx-format.md),
 #' but these can be overridden by the user.
 #'
-#' For `file_engine()`s, preferences are also used to set the node properties to search when using `search_nodes()`, defaulting to regex-based searches on id, name, and description.
-#'
 #' @param filename A character string indicating the filename or URL of the KGX-based tsv file.
 #' @param preferences A named list of preferences for the engine.
 #' @param ... Additional arguments (unused).
@@ -22,29 +20,10 @@
 #'
 #' # Using a local MONDO KGX file (packaged with monarchr)
 #' filename <- system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")
-#' e <- file_engine(filename)
-#' print(e$preferences)   # print the default preferences
+#' engine <- file_engine(filename)
 #'
-#' # same search and fetch, different preferences
-#' search_nodes(e, "fibrosis", limit = 5) |>
-#'   activate(nodes) |>
-#'   as.data.frame() |>
-#'   select(name, id, pcategory, category)
-#'
-#' # prefer to set pcategory to "biolink:ThingWithTaxon" if it applies,
-#' # followed by "biolink:NamedThing", otherwise use the first listed category.
-#' # Additionally, only search nodes' name property.
-#' # (Note that the MONDO KGX file does not provide multiple categories for
-#' # nodes, so the category_priority preference has no effect here.)
-#' e <- file_engine(filename,
-#'                  preferences = list(category_priority = c("biolink:ThingWithTaxon",
-#'                                                           "biolink:NamedThing"),
-#'                                     node_search_properties = c("name")))
-#'
-#' search_nodes(e, "fibrosis", limit = 5) |>
-#'   activate(nodes) |>
-#'   as.data.frame() |>
-#'   select(name, id, pcategory, category)
+#' res <- engine |> fetch_nodes(query_ids = c("MONDO:0007522", "MONDO:0007947"))
+#' print(res)
 #'
 #' @importFrom archive archive_read
 #' @importFrom readr col_character